fix(webui): harden drag grid review fixes
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
